+++ src/app/widgets/SOU/sou-mvp-mnemonic-scheme/components/sou-schema/sou-schema.component.ts
@@ -23,6 +23,13 @@ interface IElementFull {
     textPercent: Element;
     ellipse: Element;
     flag: boolean;
+    textParams?: ITypeTextParams;
+}
+
+interface ITypeTextParams {
+    lineLength?: number, // Макс. длина строки
+    lineHeight?: number, // Высота одной строки в пикселях
+    maxTextLength?: number, // Макс. длина текста до усечения
 }
 
 type TypeMode = 'standard' | 'deviation' | 'disabled' | 'reset' | 'active';
@@ -34,10 +41,48 @@ type TypeMode = 'standard' | 'deviation' | 'disabled' | 'reset' | 'active';
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class SouSchemaComponent implements OnChanges {
+
     elementsNode: Element[] = []; // все элементы
     elementsMap: Map<number, Element> = new Map(); // Svg элементы .element
     elementsFullMap: Map<number, IElementFull> = new Map(); // Распарсеные элементы
     dataPark: SouSectionData[] = []; // Данные с бэка
+
+    private typesNeedTextAnchorMiddle: number[] = [4, 12, 13, 14, 16];
+    private typesNeedTextAnchorEnd: number[] = [15];
+    private typeTextParams: {[typeId: number]: ITypeTextParams} = {
+        1: {
+            lineLength: 15,
+            lineHeight: 14,
+            maxTextLength: 30,
+        },
+        2: {
+            lineLength: 17,
+            lineHeight: 14,
+            maxTextLength: 17 * 2,
+        },
+        3: {
+            lineLength: 23,
+            lineHeight: 12,
+        },
+        4: {
+            lineLength: 13,
+            lineHeight: 20,
+        },
+        11: {
+            maxTextLength: 10,
+        },
+        12: {
+            maxTextLength: 7,
+        },
+        13: {
+            lineLength: 7,
+            lineHeight: 20,
+        },
+        16: {
+            maxTextLength: 17,
+        },
+    };
+    private debugElementCode: number;
 
     @Input() sectionsData: SouSectionData[];
     @Input() chosenSetting: number = 1;
@@ -202,7 +247,17 @@ export class SouSchemaComponent implements OnChanges {
     // Обработка данных с бека
     // Тут можно замокать данные
     private processSectionsData(): void {
-        this.dataPark = this.sectionsData;
+        if (this.debugElementCode) {
+            this.dataPark = this.sectionsData.map((item: SouSectionData) => {
+                if (item?.code === this.debugElementCode) {
+                    console.log(`Отладка элемента: ${item.code}. Данные с бэка:`, item);
+                }
+
+                return item;
+            });
+        } else {
+            this.dataPark = this.sectionsData;
+        }
         this.updateSvgBySectionData(false);
     }
 
@@ -224,16 +279,25 @@ export class SouSchemaComponent implements OnChanges {
 
         if (reload) {
             if (element?.children) {
-                this.addClassAndTextToElement(element, this.elementsFullMap?.get(sectionData?.code), mode);
+                const elementFull = this.elementsFullMap?.get(sectionData?.code);
+                this.addClassAndTextToElement(element, elementFull, mode);
             }
         } else {
             this.prepareElement(mode, element, sectionData);
         }
     }
 
-    private getElementMode(data: SouSectionData): TypeMode {
-        const {isExceedingConfInterval, isEnable} = data;
-        return isExceedingConfInterval ? 'deviation' : isEnable ? 'standard' : 'disabled';
+    private getElementMode(sectionData: SouSectionData): TypeMode {
+        const isMeasured = sectionData?.tolerance === 0; // Измеряемый
+        const {isExceedingConfInterval, isEnable} = sectionData;
+
+        if (isExceedingConfInterval && isEnable && isMeasured) {
+            return 'deviation';
+        } else if (!isEnable) {
+            return 'disabled';
+        }
+
+        return 'standard';
     }
 
     relatedArray(related: number[] | string): number[] {
@@ -261,11 +325,13 @@ export class SouSchemaComponent implements OnChanges {
 
         // Обработка элементов схемы
         elements?.forEach((element: Element) => {
-            const elMatch = element?.id?.match(/element-(\d+)_(\d+)/i);
-            const id = elMatch && elMatch[2] && parseInt(elMatch[2], 10);
-
+            const id = this.getElementId(element);
             this.prepareElement('reset', element);
             this.elementsMap.set(id, element);
+
+            if (this.debugElementCode && id === this.debugElementCode) {
+                console.log(`Отладка элемента: ${id}. Элемент на мнемосхеме:`,  element);
+            }
         });
 
         console.log(`Элементов и линий: ${this.elementsMap?.size}`);
@@ -277,42 +343,51 @@ export class SouSchemaComponent implements OnChanges {
         sectionData?: SouSectionData,
     ): void {
         if (element?.children) {
-            let elementFull: IElementFull = {
-                sectionData,
-                rects: [],
-                points: [],
-                arrows: [],
-                texts: [],
-                circle: null,
-                textValue: null,
-                textPercent: null,
-                ellipse: null,
-                flag: true,
-            };
-            const children = Array.from(element?.children);
-            // Search
-            children?.forEach((elem) => {
-                elementFull = this.searchElementsInElement(elem, elementFull);
-            });
-            // add class and text to element
-            this.addClassAndTextToElement(element, elementFull, mode);
+            const elementFull = this.getElementFull(element, sectionData);
 
-            // text-anchor для выравнивания текста по центру
-            if (this.doesElemNeedTextAnchor(element)) {
-                const textAnchorClassName = 'text-anchor';
-                this.addElemClass(elementFull?.textValue, [textAnchorClassName]);
-                this.addElemClass(elementFull?.textPercent, [textAnchorClassName]);
-                elementFull?.texts?.forEach((el: Element) => {
-                    this.addElemClass(el, [textAnchorClassName]);
-                });
+            if (this.debugElementCode && this.getElementId(element) === this.debugElementCode) {
+                console.log(`Отладка элемента: ${this.debugElementCode}. Заполнен IElementFull:`, elementFull);
             }
+
+            const elementId = this.getElementId(element);
+
+            if (elementFull) {
+                this.elementsFullMap.set(elementId, elementFull);
+            }
+
+            this.addClassAndTextToElement(element, elementFull, mode);
 
             // Event
             if (sectionData) {
                 this.addElementClickListener(element, elementFull);
-                this.elementsFullMap.set(sectionData?.code, elementFull);
             }
+
+            // Применение класса для текстовых нод
+            this.setElementTextNodeClassIfNeed(element);
         }
+    }
+
+    // Распарсить элемент и заполнить IElementFull
+    private getElementFull(element: Element, sectionData: SouSectionData): IElementFull {
+        let elementFull: IElementFull = {
+            sectionData,
+            rects: [],
+            points: [],
+            arrows: [],
+            texts: [],
+            circle: null,
+            textValue: null,
+            textPercent: null,
+            ellipse: null,
+            flag: true,
+        };
+        const children = Array.from(element?.children);
+
+        children?.forEach((elementChild: SVGElement) => {
+            elementFull = this.recognizeElementChild(elementChild, elementFull);
+        });
+
+        return elementFull;
     }
 
     // Добавление класса и текста для элемента
@@ -437,8 +512,8 @@ export class SouSchemaComponent implements OnChanges {
                         flag: true,
                     };
                     // Search
-                    elementsRelated?.forEach((elem) => {
-                        elementFullRelated = this.searchElementsInElement(elem, elementFullRelated);
+                    elementsRelated?.forEach((elem: SVGElement) => {
+                        elementFullRelated = this.recognizeElementChild(elem, elementFullRelated);
                     });
                     elementFullRelated?.rects?.forEach((item: Element) => {
                         this.setElementMode(item, mode);
@@ -464,8 +539,9 @@ export class SouSchemaComponent implements OnChanges {
         }
     }
 
-    // Функция распознает, чем является элемент. Если чем то нужным, то кладет его в IElementFull
-    searchElementsInElement(element: Element, elementFull: IElementFull): IElementFull {
+    // Функция распознает, потомка элемента схемы.
+    // Если он является чем то нужным, то кладет его в IElementFull
+    recognizeElementChild(element: SVGElement, elementFull: IElementFull): IElementFull {
         const name = element.getAttribute('id');
         if (name?.includes('rect')) {
             if (element?.children?.length) {
@@ -486,6 +562,11 @@ export class SouSchemaComponent implements OnChanges {
             }
         } else if (name?.includes('point')) {
             elementFull?.points.push(element);
+        } else if (name?.includes('text-params_')) {
+            const textParams = this.parseTextParamsFromElementName(name);
+            if (textParams) {
+                elementFull.textParams = textParams;
+            }
         } else if (name?.includes('text')) {
             if (name.includes('text_value')) {
                 elementFull.textValue = element;
@@ -502,6 +583,23 @@ export class SouSchemaComponent implements OnChanges {
             elementFull.ellipse = element;
         }
         return elementFull;
+    }
+
+    // Распознает индивидуальные настройки текста из id элемента
+    private parseTextParamsFromElementName(elementChildName: string): ITypeTextParams {
+        const match = elementChildName.match(/text-params_(\w+)_(\d+)/i);
+        if (match && match[1] && match[2]) {
+            const paramName = match[1];
+            const value = parseInt(match[2], 10);
+
+            if (paramName && value) {
+                return {
+                    [paramName]: value,
+                };
+            }
+        }
+
+        return null;
     }
 
     // Добавляет отслеживание клика на элементе
@@ -644,33 +742,46 @@ export class SouSchemaComponent implements OnChanges {
     }
 
     // Добавляет текст для текстовой ноды <text>
-    // @TODO Element можно заменить на SVGElement
-    private addTextToTextElem(textElem: Element, text: string): void {
+    private addTextToTextElem(textElem: SVGElement | Element, text: string): void {
         if (textElem?.children) {
             this.makeTextElemMultilineIfNeed(textElem as SVGElement, text);
             const textParams = this.getTextElemLayoutParams(textElem as SVGElement);
-
             const children = Array.from(textElem?.children);
+            let truncatedText = text;
+            const lineLength = textParams?.lineLength;
+
+            if (textParams?.maxTextLength && text?.length > textParams.maxTextLength) {
+                truncatedText = text.slice(0, textParams.maxTextLength - 3) + '...';
+                this.addTooltipToTextElem(textElem as SVGElement, text);
+            }
 
             if (text === '') {
                 children?.forEach((child: Element) => {
                     this.setTspanText(child, text);
                 });
             }
-            if (textParams?.lineLength && text.length > textParams.lineLength) {
+
+            if (lineLength && truncatedText.length > lineLength) {
                 if (children.length > 1) {
                     children.forEach((child: SVGTextPositioningElement, index: number) => {
-                        const from = textParams.lineLength * index;
-                        const to = textParams.lineLength * (index + 1);
-                        this.setTspanText(child, text.slice(from, to).trim());
+                        const from = lineLength * index;
+                        const to = lineLength * (index + 1);
+                        this.setTspanText(child, truncatedText.slice(from, to).trim());
                     });
                 } else {
-                    this.setTspanText(children[0], text);
+                    this.setTspanText(children[0], truncatedText);
                 }
             } else {
-                this.setTspanText(children[0], text);
+                this.setTspanText(children[0], truncatedText);
             }
         }
+    }
+
+    private addTooltipToTextElem(textElem: SVGElement, tooltipText: string): void {
+        const titleElem = this.renderer.createElement('title', 'http://www.w3.org/2000/svg');
+        const text = this.renderer.createText(tooltipText);
+        this.renderer.appendChild(titleElem, text);
+        this.renderer.appendChild(textElem, titleElem);
     }
 
     // Сделать текст <text> многострочным для элементов определенных типов
@@ -678,9 +789,14 @@ export class SouSchemaComponent implements OnChanges {
     private makeTextElemMultilineIfNeed(textElem: SVGElement, text: string): void {
         const textParams = this.getTextElemLayoutParams(textElem);
         const lineOffsetTopPx = textParams?.lineHeight || 20;
+        const lineLength = textParams?.lineLength;
+        const maxTextLength = textParams?.maxTextLength;
 
-        if (textParams?.lineLength) {
-            const linesCount = Math.ceil(text?.length / textParams.lineLength);
+        if (lineLength) {
+            const textLength = (maxTextLength && (text?.length > maxTextLength))
+                ? maxTextLength
+                : text?.length;
+            const linesCount = Math.ceil(textLength / lineLength);
             const children = textElem?.children && Array.from(textElem.children) as SVGTextPositioningElement[];
 
             if (children?.length < linesCount) {
@@ -705,50 +821,59 @@ export class SouSchemaComponent implements OnChanges {
 
     // Возвращает параметры перестроения текста <text>
     // Или null если ограничений нет
-    private getTextElemLayoutParams(textElem: SVGElement): {
-        lineLength: number, // Максимальное количество символов в строке
-        lineHeight: number, // Высота новой строки при переносе
-    } {
+    private getTextElemLayoutParams(textElem: SVGElement): ITypeTextParams {
         const elementTypeId = this.getElementTypeId(textElem?.parentElement);
 
-        switch (elementTypeId) {
-            case 1:
-                return {
-                    lineLength: 17,
-                    lineHeight: 14,
-                };
-            case 3:
-                return {
-                    lineLength: 23,
-                    lineHeight: 12,
-                };
-            case 4:
-                return {
-                    lineLength: 13,
-                    lineHeight: 20,
-                };
-            case 13:
-                return {
-                    lineLength: 7,
-                    lineHeight: 20,
-                };
+        if (elementTypeId) {
+            const textParams = this.typeTextParams[elementTypeId];
+            const elementId = this.getElementId(textElem?.parentElement);
+            const elementFull = this.elementsFullMap.get(elementId);
+            const individualTextParams = elementFull?.textParams;
+
+            if (individualTextParams) {
+                return Object.assign(textParams, individualTextParams);
+            }
+
+            return textParams;
         }
 
         return null;
     }
 
-    // Нужно ли этому элементу выравнивание текста по центру
-    private doesElemNeedTextAnchor(element: Element): boolean {
-        const typesNeedTextAnchor = [4, 13, 14];
-        const elementTypeId = this.getElementTypeId(element);
-        return elementTypeId && typesNeedTextAnchor.includes(elementTypeId);
+    // Возвращает ID элемента
+    private getElementId(element: SVGElement | Element): number {
+        const elMatch = element?.id?.match(/element-(\d+)_(\d+)/i);
+        return elMatch && elMatch[2] && parseInt(elMatch[2], 10);
     }
 
     // Возвращает ID типа элемента
     private getElementTypeId(element: SVGElement | Element): number {
         const elMatch = element?.id?.match(/element-(\d+)_(\d+)/i);
-        const typeId = elMatch && elMatch[1] && parseInt(elMatch[1], 10);
-        return typeId;
+        return elMatch && elMatch[1] && parseInt(elMatch[1], 10);
+    }
+
+    // Установка класса для текстовой ноды внутри элемента если нужно (для выравнивания текста)
+    private setElementTextNodeClassIfNeed(element: SVGElement | Element): void {
+        const elementTypeId = this.getElementTypeId(element);
+        let className: string;
+
+        if (elementTypeId) {
+            if (this.typesNeedTextAnchorMiddle.includes(elementTypeId)) {
+                className = 'text-anchor-middle';
+            } else if (this.typesNeedTextAnchorEnd.includes(elementTypeId)) {
+                className = 'text-anchor-end';
+            }
+        }
+
+        if (className) {
+            const elementId = this.getElementId(element);
+            const elementFull = this.elementsFullMap.get(elementId);
+            this.addElemClass(elementFull?.textValue, [className]);
+            this.addElemClass(elementFull?.textPercent, [className]);
+            elementFull?.texts?.forEach((el: Element) => {
+                this.addElemClass(el, [className]);
+            });
+        }
     }
 
     searchArrow(elements: HTMLCollection): Element[] {
@@ -832,4 +957,5 @@ export class SouSchemaComponent implements OnChanges {
         }
 
     }
+
 }